requirements.txt
colorlog==6.7.0
homeassistant>=2023.6
pip>=21.0
ruff==0.0.255

bleak
bleak-retry-connector
bluetooth-sensor-state-data